==> src/01_retrain-cadd_v1.3/01_Load_CADD_training.ipynb <==
import dask.dataframe as dd
import pandas as pd

## Data loading
First, we load the training dataset, inspect it and determine how we want to partition it.
According to the author, train.csv.gz is the *"fully expanded variant set for direct training in a learning package"*.

cadd = '/s/project/kipoi-cadd/data/raw/v1.3/training_data/training_data.tsv.gz'
cadd_df = pd.read_csv(cadd, sep='\t', compression='gzip')
cadd_df.head()

## Replication of results
- Aiming for CADD v1.3 which was trained using the Logistic Regression module from GraphLab Create v1.4
- From their release notses: "(..) we trained only one classifier using approximately 15 million human derived variants (newly extracted from EPO 6 primate alignments v75) versus approximately 15 million (newly) simulated variants. We used an L2 penalty of 1.0 and terminated training after 10 iterations."
