==> src/01_retrain-cadd_v1.3/02_Separating_per_chromosome.ipynb <==
## Saving each chromosome as a separate parquet file

import dask.dataframe as dd
import pandas as pd
from fastparquet import write
from fastparquet import ParquetFile
from m_kipoi.config import VCF_HEADER  # hg19 based
from collections import OrderedDict

all_variants = '/s/project/kipoi-cadd/data/processed/kipoi/all_variants'
parquet = dd.read_parquet(all_variants, index="ID")
parquet.head()

chrom = 2
temp = parquet[parquet.Chrom == str(chrom)]
chrom_df = temp.compute()
chrom_df.Chrom = chrom_df.Chrom.astype('int64')

chrom_parquet = '/s/project/kipoi-cadd/data/processed/kipoi/chr' + str(chrom) + '/variant_labels.parq'
write(chrom_parquet, chrom_df)

### Checking that it worked...

pf = ParquetFile('/s/project/kipoi-cadd/data/processed/kipoi/chr1/variant_labels.parq')
df = pf.to_pandas()
df.head()

## Saving each chromosome as VCF file
Using some of the tools provided by Kipoi to manage VCF files.
1. Pandas -> VCF file: [Kipoi code](https://github.com/kipoi/manuscript/blob/master/src/splicing/dbscSNV.smk#L44-L65)
2. BZip and tabix the resulting VCF file: [Kipoi code](https://github.com/kipoi/manuscript/blob/master/src/splicing/dbscSNV.smk#L75-L77)

chrom = 22
pf = ParquetFile('/s/project/kipoi-cadd/data/processed/kipoi/chr' + str(chrom) + '/variant_labels.parq')
df = pf.to_pandas()
df.sort_values(by='Pos', inplace=True)
df.head()

# Write the header
output = '/s/project/kipoi-cadd/data/processed/kipoi/chr' + str(chrom) + '/variants.vcf'

with open(output, "w") as f:
    f.write(VCF_HEADER)

# Append the variants
pd.DataFrame(OrderedDict([("#CHROM", df.Chrom.astype(str)),
                          ("POS", df.Pos),
                          ("ID", df.index.values),
                          ("REF", df.Ref),
                          ("ALT", df.Alt),
                          ("QUAL", "."),
                          ("FILTER", "."),
                          ("INFO", "."),
                          ])).to_csv(output, mode='a', header=True, index=False, sep="\t")
