==> src/01_retrain-cadd_v1.3/03_Running_kipoi_veff.ipynb <==
## Running kipoi_veff in for one chromosome
Trying to run DeepBind/Homo_sapiens/RBP/D00084.001_RNAcompete_A1DF. The [documentation](http://kipoi.org/models/DeepBind/Homo_sapiens/RBP/D00084.001_RNAcompete_A1CF/) says:
```
cd ~/.kipoi/models/DeepBind/Homo_sapiens/RBP/D00084.001_RNAcompete_A1CF
kipoi predict DeepBind/Homo_sapiens/RBP/D00084.001_RNAcompete_A1CF \
  --dataloader_args='{'intervals_file': 'example_files/intervals.bed', 'fasta_file': 'example_files/hg38_chr22.fa'}' \
  -o '/tmp/DeepBind|Homo_sapiens|RBP|D00084.001_RNAcompete_A1CF.example_pred.tsv'
'# check the results
head '/tmp/DeepBind|Homo_sapiens|RBP|D00084.001_RNAcompete_A1CF.example_pred.tsv'
```

from fastparquet import ParquetFile
import pandas as pd

chrom = 22
pf = ParquetFile('/s/project/kipoi-cadd/data/processed/kipoi/chr'+str(chrom)+'/variant_labels.parq')
df = pf.to_pandas()
df.head()

intervals_df = df.rename(columns={'Chr':'chr', 'Pos': 'start'})
intervals_df['end'] = intervals_df.start + 1
intervals_df['strand'] = ['*'] * intervals_df.shape[0]


intervals = '/s/project/kipoi-cadd/data/processed/kipoi/chr'+str(chrom)+'/intervals.tsv'
intervals_df.to_csv(intervals, sep='\t', ind)
