==> src/01_retrain-cadd_v1.3/06_Test_lmdb.ipynb <==
## Test put/get from lmdb
### ouga04

from kipoi_cadd.tests.test_lmdb import test_get_10000_variants, test_put_10000_variants

%%time
test_put_10000_variants()

%%time
test_get_10000_variants()

# In hours
put_whole_dataset = ((34000000/10000)*6.42)/3600 # 6.42 seconds per 10000
# In seconds
get_whole_dataset = ((34000000/10000)*6.36)/1000 # 6.36 ms per 10000
print("Put: {:0.3f}".format(put_whole_dataset), "hours", "\nGet:", get_whole_dataset, "seconds")

## gpu01

%%time
test_put_10000_variants()

%%time
test_get_10000_variants()

# In hours
put_whole_dataset = ((34000000/10000)*5.36)/3600 # 5.36 seconds per 10000
# In seconds
get_whole_dataset = ((34000000/10000)*81.2)/1000/60 # 81.2 ms per 10000
print("Put: {:0.3f}".format(put_whole_dataset), "hours", "\nGet: {:0.3f}".format(get_whole_dataset), "minutes")

(34000000/10000)*44160000

83904000/1.9

150*1.9
