==> src/01_retrain-cadd_v1.3/07_Test_dataloader.ipynb <==
## Test CaddDataset

from gin_train.trainers import KerasTrainer

from kipoi_cadd.data import CaddDataset
from kipoi_cadd.data import cadd_train_valid_data
from kipoi_cadd.data import save_variant_ids
from kipoi_cadd.data import train_test_split_indexes
from kipoi_cadd.config import get_data_dir
from random import shuffle
from kipoi_cadd.models import logistic_regression_scikit, logistic_regression_keras
from gin_train.metrics import ClassificationMetrics
from sklearn.model_selection import train_test_split
%load_ext line_profiler

## 1. Check the dataset class

variant_id_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/sample_variant_ids.pkl"
lmdb_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/lmdb/"
ds = CaddDataset(lmdb_dir, variant_id_file)
print("Version", ds.version)
ds.__getitem__(234)

set_10000 = list(range(10000))
shuffle(set_10000)

%%time
for var in set_10000:
    ds.__getitem__(var)

%lprun -f ds.__getitem__ for var in range(512): ds.__getitem__(var)

%%time
# Create train, test and validation indexes
all_variant_id_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/variant_ids.pkl"
training_data_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/"
train_idx_path = training_data_dir + "train_idx.pkl"
test_idx_path = training_data_dir + "test_idx.pkl"
valid_idx_path = training_data_dir + "valid_idx.pkl"
lmdb_dir = training_data_dir + "lmdb"
"""
train_idx, test_idx = train_test_split_indexes(all_variant_id_file, test_size=0.4)
save_variant_ids(test_idx_path, test_idx)
test_idx, valid_idx = train_test_split_indexes(test_idx_path, test_size=0.5)

save_variant_ids(train_idx_path, train_idx)
save_variant_ids(test_idx_path, test_idx)
save_variant_ids(valid_idx_path, test_idx)
"""
train, valid = cadd_train_valid_data(lmdb_dir, train_idx_path, valid_idx_path)

%%time
train_it = train.batch_train_iter(batch_size=512, num_workers=10, shuffle=True)

## TODO

- can we have multiple connections to LMDB open at the same time?
  - just open two lmdb's in the same python or different python session

%%time
ds = CaddDataset(lmdb_dir, train_idx_path)
ds_iter = ds.batch_train_iter(batch_size=512, num_workers=1, shuffle=True)
next(ds_iter)

## 2. Train a logistic regression model

tr = KerasTrainer(logistic_regression_keras(n_features=1063), train, valid, get_data_dir() + "/models/try10")
tr.train(batch_size=512, epochs=10, num_workers=10)

metric = ClassificationMetrics()
tr.evaluate(metric)

# training
train.batch_size = 256  # 512
train.num_workers = 10

# train-specific
train.epochs = 200
train.early_stop_patience = 4
train.train_epoch_frac = 1.0  # optionally subset the training set
train.valid_epoch_frac = 1.0  # optionally subset the validation set
