==> src/01_retrain-cadd_v1.3/08_Test_load_db_in_memory.ipynb <==
## Test load database in memory

from kipoi_cadd.data import cadd_train_valid_data, train_test_split_indexes
from kipoi_cadd.utils import dump_to_pickle

lmdb_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/lmdb"
variant_id_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/variant_ids.pkl"
train_id_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/train_ids.pkl"
valid_id_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/valid_ids.pkl"

train_idx, valid_idx = train_test_split_indexes(variant_id_file, 0.2, random_state=1)
dump_to_pickle(train_id_file, train_idx)
dump_to_pickle(valid_id_file, valid_idx)

train, valid = cadd_train_valid_data(lmdb_dir, train_id_file, valid_id_file)

X_train, y_train = train.load_all(64, 2)
X_valid, y_valid = valid.load_all(64, 2)

# Fixing test data that has only one class
from random import randint

idx = [randint(1,1000) for x in range(20)]
y_train[idx] = 1
y_valid[idx] = 1

## Test training with sklearn

from kipoi_cadd.models import logistic_regression_scikit

tr = logistic_regression_scikit()
tr.fit(X, y)

tr.score(X_valid, y_valid)
