==> src/01_retrain-cadd_v1.3/11_Summary_statistics.ipynb <==
## Summary statistics

from kipoi_cadd.data import cadd_train_valid_data, train_test_split_indexes, CaddDataset
from kipoi_cadd.utils import dump_to_pickle, load_pickle
import pandas as pd
import time
from tqdm import tqdm, trange
import numpy as np
from kipoi.data_utils import numpy_collate_concat
from sklearn.preprocessing import StandardScaler
%load_ext line_profiler

training_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/"
lmdb_dir = training_dir + "lmdb"
csv_file = training_dir + "training_data.imputed.csv"
valid_id_file = training_dir + "valid_idx.pkl"
all_ids_file = training_dir + "variant_ids.pkl"
ids_10k_file = training_dir + "sample_indices/ids_10k.pkl"
first_10k_file = training_dir + "sample_indices/first_10k_ids.pkl"
last_10k_file = training_dir + "sample_indices/last_10k_ids.pkl"
last_10k_csv = training_dir + "last10k.csv"

## Fist 10k rows of the csv file

df10k = pd.read_csv(csv_file, nrows=10000)
df10k_y = [0 if s == -1 else 1 for s in df10k.y]
df10k.drop('y', axis=1, inplace=True)
pd_description = df10k.describe(include='all')
dump_to_pickle(training_dir + "stats/summary_stats_first10k.pkl", pd_description)
pd_description

"""
all_ids = load_pickle(all_ids_file)
first_10k = all_ids.iloc[:10000]
first_10k_file = training_dir + "first_10k_ids.pkl"
dump_to_pickle(first_10k_file, first_10k)
"""

first_10k_ds = CaddDataset(lmdb_dir, first_10k_file)
X, y = first_10k_ds.load_all(shuffle=False, drop_last=False)

np.allclose(np.array(df10k, dtype=np.float16), X, rtol=1e-05, atol=1e-07)

np.allclose(np.array(df10k_y, dtype=np.float16), y, rtol=1e-05, atol=1e-07)

## Last 10k rows of the csv file

"""
! head {csv_file} -n 1 > last10k.csv
! tail {csv_file} -n 10000 >> last10k.csv
"""

df_last10k = pd.read_csv(last_10k_csv)
df_last10k_y = [0 if s == -1 else 1 for s in df_last10k.y]
df_last10k.drop('y', axis=1, inplace=True)
pd_description_last10k = df_last10k.describe(include='all')
dump_to_pickle(training_dir + "stats/summary_stats_last10k.pkl", pd_description_last10k)
pd_description_last10k

"""
all_ids = load_pickle(all_ids_file)
last_10k = all_ids.tail(10000)
dump_to_pickle(last_10k_file, last_10k)
"""
last_10k = load_pickle(last_10k_file)

last_10k_ds = CaddDataset(lmdb_dir, last_10k_file)
X, y = last_10k_ds.load_all(shuffle=False, drop_last=False)

ds = CaddDataset(lmdb_dir, last_10k_file)
full_info = ds.load_all_with_metadata(shuffle=False, drop_last=False)

equality = np.array(df_last10k, dtype=np.float16) == X

dif_idx = {}
for i in range(10000):
    for j in range(1063):
        if not equality[i,j]:
            if i in dif_idx.keys():
                dif_idx[i].append(j)
            else:
                dif_idx[i] = []

len([v if "N" in v else "" for v in full_info['metadata']['variant_id']])

for k, v in dif_idx.items():
    print(len(v))

# dif_idx has only these rows 915, 7008, 9223
pd.DataFrame([np.array(df_last10k, dtype=np.float16)[915], full_info['inputs'][915]]).iloc[:,10:50]

full_info['inputs'][915]

np.allclose(np.array(df_last10k, dtype=np.float16), full_info['inputs'], rtol=1e-05, atol=1e-07)

np.allclose(np.array(df_last10k_y, dtype=np.float16), full_info['targets'], rtol=1e-05, atol=1e-07)

## Standardize dataset using StandardScaler

first_10k_ds = CaddDataset(lmdb_dir, first_10k_file)
X, y = first_10k_ds.load_all(shuffle=False, drop_last=False)

scaler = StandardScaler(with_mean=False) # Used with_mean=False, for it to work with sparse matrices too.
X = X.astype(np.float32)
X_transformed = scaler.fit_transform(X)

pd.DataFrame(X).describe(include='all')

pd.DataFrame(X_transformed).describe(include='all').loc[:,124]

X_small = X_transformed.astype(np.float16)
pd.DataFrame(X_small).describe(include='all').loc[:,124]

# Save scaler for later
dump_to_pickle(training_dir + "stats/standard_scaler_first10k.pkl", scaler)

## Standardize dataset using MinMaxScaler

first_10k_ds = CaddDataset(lmdb_dir, first_10k_file)
X, y = first_10k_ds.load_all(shuffle=False, drop_last=False)

scaler = StandardScaler(with_mean=False) # Used with_mean=False, for it to work with sparse matrices too.
X = X.astype(np.float32)
X_transformed = scaler.fit_transform(X)
print(scaler.mean_)

# Save scaler for later
dump_to_pickle(training_dir + "stats/minmax_scaler_first10k.pkl", scaler)

recovered = load_pickle(training_dir + "stats/minmax_scaler_first10k.pkl")
recovered.transform(np.ones_like(X))

recovered.mean_
