==> src/01_retrain-cadd_v1.3/12_Test_sparse_matrices.ipynb <==
from kipoi_cadd.data_utils import load_csv_to_sparse_matrix
from kipoi_cadd.data import cadd_train_valid_data, train_test_split_indexes, CaddDataset
from kipoi_cadd.utils import dump_to_pickle, load_pickle
import pandas as pd
import time
from tqdm import tqdm, trange
import numpy as np
from kipoi.data_utils import numpy_collate_concat
from scipy.sparse import csr_matrix, save_npz, load_npz, vstack
%load_ext line_profiler

training_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/"
lmdb_dir = training_dir + "lmdb"
csv_file = training_dir + "training_data.imputed.csv"
valid_id_file = training_dir + "valid_idx.pkl"
all_ids_file = training_dir + "variant_ids.pkl"
ids_10k_file = training_dir + "sample_indices/ids_10k.pkl"
std_scaler_file = training_dir + "stats/standard_scaler_first10k.pkl"
short_csv = training_dir + "last10k.csv"
sparse_all = training_dir +  "sparse_all.npz"

## a. Use read_csv with file buffer and nrows
Here we realize that `nrows` is probably not meant to be used together with buffers. We obtain weird results.

f = open(csv_file, "rb")

df = pd.read_csv(f, nrows=1000)
header = df.columns.values
dtypes = df.dtypes.to_dict()
df.head()

df2 = pd.read_csv(f, nrows=1000, names=header)
df2.head()

f.close()

df3 = pd.read_csv(csv_file, nrows=200)
df3.loc[99:105,:]

## b. Load entire sparse csv directly into a sparse matrix using Dask
I saw this approach here https://stackoverflow.com/questions/31888856/read-a-large-csv-into-a-sparse-pandas-dataframe-in-a-memory-efficient-way.
After loading the sparse matrix, we want to **shuffle** it and then **persist** it. The dataset size in disk should be considerably reduced.

colnames = pd.read_csv(short_csv, nrows=1).columns.values
dic = {}
for i, col in enumerate(colnames):
    dic[col] = i

%%time
# Load sparse matrix directly from csv file
res = load_csv_to_sparse_matrix(short_csv)

# Shuffle a sparse matrix
index = load_pickle(training_dir + "shuffle_splits/shuffled_index_10k.pkl")
print(index)
shuff_res = res[index, :]
shuff_res

print(shuff_res[:,0].shape)
print(shuff_res[:,1:].shape)

# Persist csr_matrix
save_npz(training_dir + "example_sparse.npz", shuff_res)

## c. Load chunks of csv file using pandas (WINNER)
Let's try out here with a smaller csv file and concatenate the sparse matrices.

def load_csv_chunks_tosparse(filename, chunksize, dtype, num_lines, output=None, header=0, index_col=None):
    full_matrix, chunk_nr = None, 0
    tqdm_total = num_lines//chunksize if num_lines % chunksize > 0 else (num_lines//chunksize) - 1
    
    for chunk in tqdm(pd.read_csv(filename, chunksize=chunksize, dtype=dtype, header=header,
                                  index_col=index_col), total=tqdm_total):
        if full_matrix is None:
            full_matrix = csr_matrix(chunk, shape=chunk.shape, dtype=dtype)
        else:
            chunk_csr = csr_matrix(chunk, shape=chunk.shape, dtype=dtype)
            full_matrix = vstack([full_matrix, chunk_csr])
    
        chunk_nr += 1
        if chunk_nr >= 3: break
    if output is None:
        return full_matrix
    else:
        save_npz(output, full_matrix)

res = load_csv_chunks_tosparse(csv_file, 10000, np.float32, 34000000)

# Self contained script to create full cadd dataset sparse matrix
from kipoi_cadd.data_utils import load_csv_chunks_tosparse
from kipoi_cadd.utils import load_pickle
import numpy as np

lmdb_dir = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/lmdb"
csv_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/training_data.imputed.csv"
all_ids_file = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/variant_ids.pkl"
output = "/s/project/kipoi-cadd/data/raw/v1.3/training_data/sparse_all_ordered.npz"

all_ids = load_pickle(all_ids_file)
num_lines = len(all_ids)
del all_ids
load_csv_chunks_tosparse(csv_file, 50000, np.float32, num_lines, output=output, header=0)

19999 % 10000

res

print(whole_thing.count_nonzero()/(whole_thing.shape[0]*whole_thing.shape[1]))

## Extract training and validation set from sparse matrix in disk

from sklearn.model_selection import train_test_split
def sparse_cadd_datset(sparse_matrix_file, split=0.3, random_state=42):
    sparse_matrix = load_npz(sparse_matrix_file)
    y = sparse_matrix[:,0]
    X = sparse_matrix[:,1:]
    del sparse_matrix

    X_train, X_valid, y_train, y_valid = train_test_split(X, y, test_size=split, random_state=random_state)
    return (X_train, y_train), (X_valid, y_valid)

(X_train, y_train), (X_valid, y_valid) = sparse_cadd_datset(training_dir + "example_sparse.npz")

print("Finished loading training dataset. Shape: ", X_train.shape, "True values:", y_train.sum()/y_train.shape[0])

y_train.shape[0]

## Verify order of variants in sparse matrix
The original csv was loaded into sparse format using delayed functions of dask. This probably disrupted the order of variants so we want to verify if this is the case.

# Load first 10k rows of sparse matrix
tmp = load_npz(sparse_all)
sparse_10k = tmp[:10000, :]
sparse_10k = sparse_10k.toarray()
del tmp

# Load first 10k rows of csv
tmp = pd.read_csv(csv_file, nrows=10000)
dense_10k = tmp.values
del tmp

# Compare
sparse_10k.dtype == dense_10k.dtype
np.unique(sparse_10k[:, 1:] == dense_10k[:, 1:])

# THEY ARE NOT EQUAL!!!   D':
Fix this

"""
# Add idx column to dataset
awk -F'\t' -v OFS=',' '
  NR == 1 {print "ID", $0; next}
  {print (NR-2), $0}
' training_data.imputed.csv > /s/project/kipoi-cadd/data/processed/v1.3/training_data/training_data.imputed.index.csv

awk -F'\t' -v OFS=',' '
    NR == 1 {print "ID", $0; next}
    {print (NR-2), $0}
    ' shuffled_10k.csv > /s/project/kipoi-cadd/data/processed/v1.3/training_data/shuffled_10k.index.csv
"""

i = 0
for chunk in pd.read_csv(short_csv, iterator=True, chunksize=10000):
    print(chunk)
    i += 1
    if i >= 2: break

## Sparse matrices transformations with scikit scalers

X, y = CaddDataset(lmdb_dir, ids_10k_file).load_all(drop_last=False)
X_sparse = csr_matrix(X, shape=None, dtype=np.float32, copy=False)

scaler = load_pickle(std_scaler_file)

X_sp_tr = scaler.transform(X_sparse)

X_sp_tr

save_npz(training_dir + "sparse_matrix_10percent.npz", X_sparse)
# save_npz("/s/project/kipoi-cadd/data/raw/v1.3/training_data/sparse_all.npz", res)
